Moved <Quiz> & <EnterSettings> into their own files
